COVID-19_Project/requirements.txt
pandas
numpy
seaborn
voila
ipyvuetify
plotly
datetime
tabula
requests
lxml
